Added routes in the index.php

+++ index.php
@@ -10,8 +10,10 @@ Database::connect();
 
 $router = new Router();
 
-$router->get('/api/v1/users', [UserControllers::class, 'getAllUsers']);
-$router->post('/api/v1/users', [UserControllers::class, 'addUser']);
-$router->get('/api/v1/users/getuser', [UserControllers::class, 'getUser']);
-$router->update('/api/v1/users/updateuser', [UserControllers::class, 'updateUser']);
-$router->delete('/api/v1/users/deleteuser', [UserControllers::class, 'deleteUser']);
+if (str_starts_with($_SERVER['REQUEST_URI'], '/api/v1/users')) {
+    $router->get('/api/v1/users', [UserControllers::class, 'getAllUsers']);
+    $router->post('/api/v1/users', [UserControllers::class, 'addUser']);
+    $router->get('/api/v1/users/getuser', [UserControllers::class, 'getUser']);
+    $router->update('/api/v1/users/updateuser', [UserControllers::class, 'updateUser']);
+    $router->delete('/api/v1/users/deleteuser', [UserControllers::class, 'deleteUser']);
+}
